ProjectileState와 RigidState 잘못 쓴거 돌려놓기 및 연결
